Refactor : DispatchUserWrapper 컴포넌트 생성 및 Header 컴포넌트에서 해당 로직 삭제
